Changed service classes to implement interfaces
