+++ backend/src/main/java/com/coders/travelx/service/UserService.java
@@ -1,34 +1,7 @@
 package com.coders.travelx.service;
 
-import com.coders.travelx.model.User;
-import com.coders.travelx.repository.UserRepository;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
+public interface UserService {
 
-@Service
-public class UserService {
+    void changePassword(String userEmail, String currentPassword, String newPassword);
 
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
-        this.userRepository = userRepository;
-        this.passwordEncoder = passwordEncoder;
-    }
-
-    public void changePassword(String userEmail, String currentPassword, String newPassword) {
-        User user = userRepository.findByEmail(userEmail)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-
-        if (!passwordEncoder.matches(currentPassword, user.getPassword())) {
-            throw new IllegalArgumentException("Current password is incorrect");
-        }
-
-        String encodedPassword = passwordEncoder.encode(newPassword);
-        user.setPassword(encodedPassword);
-
-        userRepository.save(user);
-    }
 }
